rtio: privatize APIs/symbols in a source file
This commit moves all the most of the static inline functions in rtio.h
to a private source file, as not all of the APIs/symbols needs to be
in a public header.
