+++ subsys/rtio/CMakeLists.txt
@@ -2,16 +2,17 @@
 # SPDX-License-Identifier: Apache-2.0
 
 if(CONFIG_RTIO)
-  zephyr_syscall_header(${ZEPHYR_BASE}/include/zephyr/rtio/rtio.h)
+ zephyr_syscall_header(${ZEPHYR_BASE}/include/zephyr/rtio/rtio.h)
 
-  zephyr_library()
+ zephyr_library()
 
-  zephyr_include_directories(${ZEPHYR_BASE}/subsys/rtio)
+ zephyr_include_directories(${ZEPHYR_BASE}/subsys/rtio)
 
-  zephyr_library_sources(rtio_executor.c)
-  zephyr_library_sources(rtio_init.c)
-  zephyr_library_sources(rtio_sched.c)
-  zephyr_library_sources_ifdef(CONFIG_USERSPACE rtio_syscalls.c)
+ zephyr_library_sources(rtio.c)
+ zephyr_library_sources(rtio_executor.c)
+ zephyr_library_sources(rtio_init.c)
+ zephyr_library_sources(rtio_sched.c)
+ zephyr_library_sources_ifdef(CONFIG_USERSPACE rtio_syscalls.c)
 endif()
 
 zephyr_library_sources_ifdef(CONFIG_RTIO_WORKQ rtio_workq.c)
